stable_diffusion_text_inpaint/requirements.txt
diffusers>=0.24.0
transformers>=4.36.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
opencv-python>=4.8.0
matplotlib>=3.7.0
click>=8.1.0 